Variable message lenghts: finish feature
Messages can now be sent with individual message lengths for each
participant's message.
Messages still have to be dividable by 4, so their length will be padded
with 0s up to the next suitable lenght.
The reason for that is that fail-stop-key generation need to do some
assumptions about message lenghts.

E.g. a 16 byte message can stay 16 byte, whereas a 17 byte message will
be sent as a 20 byte message.